Fix all my cockups after rebase on master

+++ src/lib/OTA/OTA.cpp
@@ -587,14 +587,14 @@ void OtaPackAirportData(OTA_Packet_s * const otaPktPtr, FIFO<AP_MAX_BUF_LEN> *in
     uint8_t count = inputBuffer->size();
     if (OtaIsFullRes)
     {
-        count = std::min(count, (uint16_t)ELRS8_TELEMETRY_BYTES_PER_CALL);
+        count = std::min(count, (uint8_t)ELRS8_TELEMETRY_BYTES_PER_CALL);
         otaPktPtr->full.airport.count = count;
         otaPktPtr->full.airport.tlmFlag = tlmFlag;
         inputBuffer->popBytes(otaPktPtr->full.airport.payload, count);
     }
     else
     {
-        count = std::min(count, (uint16_t)ELRS4_TELEMETRY_BYTES_PER_CALL);
+        count = std::min(count, (uint8_t)ELRS4_TELEMETRY_BYTES_PER_CALL);
         otaPktPtr->std.airport.count = count;
         otaPktPtr->std.airport.tlmFlag = tlmFlag;
         inputBuffer->popBytes(otaPktPtr->std.airport.payload, count);
